ListBox: add new on_selected signal
The first item of the list is selected by default. Thus, when there is a
click on it, nothing happens as the item selected has not changed. Now a
separate event can be used to see any time an item is selected, which is
different than on_selected_changed signal.

+++ src/list.cpp
@@ -132,9 +132,9 @@ void ListBox::handle(Event& event)
 
 void ListBox::selected(size_t index)
 {
-    if (m_selected != static_cast<ssize_t>(index))
+    if (index < m_sizer.count_children())
     {
-        if (index < m_sizer.count_children())
+        if (m_selected != static_cast<ssize_t>(index))
         {
             if (static_cast<ssize_t>(m_sizer.count_children()) > m_selected)
                 m_sizer.child_at(m_selected)->checked(false);
@@ -145,12 +145,18 @@ void ListBox::selected(size_t index)
             damage();
             on_selected_changed.invoke();
         }
+
+        on_selected.invoke(index);
     }
 }
 
 void ListBox::clear()
 {
-    m_sizer.remove_all();
+    if (m_sizer.count_children())
+    {
+        m_sizer.remove_all();
+        on_items_changed.invoke();
+    }
 }
 
 }
